requirements.txt
aiogram>=3.7
aiohttp>=3.9
python-dotenv>=1.0
yt-dlp
